+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add validation for the `claimingInterval` field. Refs UIREC-308.
 * View piece status change log. Refs UIREC-305.
 * Add CSV export options for new piece statuses. Refs UIREC-306.
+* Add the "Acquisition units" protected field to the receiving title form. Refs UIREC-295.
 
 ## [4.0.0](https://github.com/folio-org/ui-receiving/tree/v4.0.0) (2023-10-12)
 [Full Changelog](https://github.com/folio-org/ui-receiving/compare/v3.0.0...v4.0.0)